requirements.txt
# DisasterVision AI — minimal deps, Python 3.8+
# Use a venv: python -m venv .venv && .venv\Scripts\activate (Windows) or source .venv/bin/activate (Linux/Mac)
# For CUDA GPU use requirements-cuda.txt instead (PyTorch cu121).
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
opencv-python-headless>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
